Move compute_native_function_declaration to its own dest module
I'm planning to turn it into a class, so let's move it over first.


[ghstack-poisoned]
