+++ tools/codegen/dest/__init__.py
@@ -1,1 +1,2 @@
 from .register_dispatch_key import RegisterDispatchKey as RegisterDispatchKey
+from .native_functions import compute_native_function_declaration as compute_native_function_declaration
